% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lightcastapi_status}
\alias{lightcastapi_status}
\title{Get Lightcast JPA API Status}
\usage{
lightcastapi_status()
}
\value{
If the HTTP request succeeds, and the status code is ok (e.g. 200), an HTTP response: an S3
list with class \code{httr2_response}.
}
\description{
Quickly check the current status of the Lightcast Job Postings API.
}
\details{
This function wraps a basic request to the API's "Status" endpoint, which returns details
regarding it's current health. This function is also used to check the current request rate limit
when starting a new session, or after waiting to let the limit reset.
}
\examples{
\dontrun{
resp <- lightcastapi_status()

httr2::resp_status(resp)
# retrieves numeric HTTP status code

httr2::resp_status_desc(resp)
# retrieves a brief textual description of status
}
}
