% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightpostAPI-package.R
\docType{package}
\name{lightpostAPI-package}
\alias{lightpostAPI}
\alias{lightpostAPI-package}
\title{lightpostAPI: Easily Query the Lightcast US Job Postings API}
\description{
Allows users to more easily query the Lightcast US Job Postings API via R and returns tabular data in a `tibble` format.
}
\author{
\strong{Maintainer}: Tucker Plumlee \email{tplumlee@crec.net}

}
\keyword{internal}
