% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posting_facets.R
\name{posting_facets}
\alias{posting_facets}
\title{Request Faceted Job Postings Data from Lightcast's "Rankings" API Endpoints}
\usage{
posting_facets(
  facet,
  ...,
  status = "active",
  start = NULL,
  end = NULL,
  rank_by = "unique_postings",
  rank_limit = 10,
  min_postings = 1,
  metrics = NULL,
  nested_facet = NULL,
  nested_rank_by = NULL,
  nested_rank_limit = NULL,
  nested_min_postings = NULL
)
}
\arguments{
\item{facet}{A variable to use in faceting (i.e., diaggregating) the job
postings data being queried. This will produce an aggregate measure,
specified by \code{rank_by} (e.g., total unique postings), for each category
of the faceting variable. Categories will be returned in descending
order of the \code{rank_by} metric up to the number of categories/items
specified by \code{rank_limit}.}

\item{...}{Named arguments indicating a
filter name and appropriate values to filter on. For example
\code{soc5 = c("11-1011", "11-1021")} or \code{msa = "47900"}. For a listing of
available filters and required values/data types, see [BUILT-IN DATA
NOT YET IMPLEMENTED].}

\item{status}{Query job postings based on their status. Defaults to \code{"active"}.
Without any arguments provided to \code{start} or \code{end} (the default) this queries
currently active postings (as of the \code{latest_day} of available). If \code{start}
and \code{end} are specified, this queries postings active at any point during
that period. Can also take values of \code{"posted"} or \code{"expired"} (in which
case\code{start} and \code{end} values must also be declared).}

\item{start, end}{Start and end dates of the period to be queried. Must both
be specified and be formatted as strings in the "YYYY-MM" format. For example,
\code{start = "2025-01"} and \code{end = "2025-12"} would return results for postings
active, posted, or expired (depending on the value of \code{status}) in the
2025 calendar year.}

\item{rank_by}{A string indicating which metric to use when ranking \code{facet}
categories. Defaults to \code{"unique_postings"}. For a listing of available
metrics, see [BUILT-IN DATA NOT YET IMPLEMENTED].}

\item{rank_limit}{An integer between \code{0} and \code{1000} specifying the number of
facet categories to return, based on \code{rank_by} metric. Defaults to \code{10}.
All facet categories can be returned by specifying \code{0}, except when
faceting by skills or certifications.}

\item{min_postings}{Requires faceted categories to have at least this many
unique postings depending on the filters applied. Defaults to \code{1} for all
\code{rank_by} metrics except \code{"significance"}, which defaults to \code{3}.}

\item{metrics}{Any additonal metrics to aggregate for the faceted categories.
For a listing of available metrics, see [BUILT-IN DATA NOT YET IMPLEMENTED].}

\item{nested_facet}{A variable to use in faceting/disaggregating each of the
\code{facet} variable's categories. For example, if \code{"msa"} is provided as an
argument for \code{facet}, providing \code{"soc5"} as an argument would return
total unique postings for each SOC occupation within each MSA.}

\item{nested_rank_by}{A string indicating which metric to use when ranking
\code{nested_facet} categories. Defaults to \code{"unique_postings"}. For a listing
of available metrics, see [BUILT-IN DATA NOT YET IMPLEMENTED].}

\item{nested_rank_limit}{An integer between \code{0} and \code{100} specifying the number of
facet categories to return, based on \code{nested_rank_by} metric. Defaults to \code{10}.
All facet categories can be returned by specifying \code{0}, except when
faceting by skills or certifications.}

\item{nested_min_postings}{Requires faceted categories to have at least this many
unique postings depending on the filters applied. Defaults to \code{1} for all
\code{nested_rank_by} metrics except \code{"significance"}, which defaults to \code{3}.}
}
\value{
A tibble containing job posting data measures.
}
\description{
This function allows the user to request job postings data that
is faceted (or disaggregated) by either a single variable or by two (i.e.,
"nested" facets). This includes aggregated measures as well as timeseries
measures. The user must specify a time period and posting status to query
(e.g., all postings that were newly posted during the specified time frame
vs. those that were active at some point during the time frame), variable(s)
to use in faceting, preferences for how facets are ranked, and any filters.
}
\examples{
# COMING SOON

}
