% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_api_creds}
\alias{set_api_creds}
\title{Install Lightcast API Credentials in Your \code{.Renviron} File for Repeated Use}
\usage{
set_api_creds(id, secret)
}
\arguments{
\item{id}{string, Lightcast API client_id.}

\item{secret}{string, Lightcast API client_secret.}
}
\description{
This function will add Lightcast API credentials to your \code{.Renviron}
file so that they can be used to authenticate requests without being hardcoded.
THESE API CREDENTIALS SHOULD NOT BE SHARED OUTSIDE THE ORGANIZATION.
}
\details{
The client_id is stored using the environmental variable \code{LIGHTCAST_API_ID} and the
client_secret is stored using the variable LIGHTCAST_API_SECRET. These can be accessed
at any time by using \code{Sys.getenv("LIGHTCAST_API_ID")} or by using
\code{Sys.getenv(LIGHTCAST_API_SECRET)}. To access and return both as a character vector,
use \code{Sys.getenv(c("LIGHTCAST_API_ID", "LIGHTCAST_API_SECRET"))} (see examples).
}
\examples{

\dontrun{
set_api_creds(id = "orgname", secret = "abcd")
# Check that credentials have been added with:
Sys.getenv(c("LIGHTCAST_API_ID", "LIGHTCAST_API_SECRET"))
}
}
