% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{authenticate_api}
\alias{authenticate_api}
\title{Authorize Access to Lightcast API}
\usage{
authenticate_api()
}
\value{
An OAuth client: An S3 list with class \code{httr2_oauth_client}.
}
\description{
This function uses stored API credentials to create a client
containing the bearer token necessary to authorize any requests to the Lightcast
API.
}
\details{
NOTE: ensure that your Lightcast API credentials have been
added to your \code{.Renviron} by using the built-in \code{set_api_creds}
function prior to using this function and making requests.

In order to use the Lightcast API, a \code{Bearer} token must
accompany every request in order to provide authentication. This function
requests a token via an OAuth 2.0 Client Credentials flow using API
credentials previously stored in \code{.Renviron} as \code{LIGHTCAST_API_ID}
and \code{LIGHTCAST_API_SECRET} and returns a \code{client} object
generated using \code{httr2::oauth_client}. This \code{client} object, which
contains the \code{Bearer} token, is attached to each API request via
\code{httr2::req_oath_client_credentials()}. This process benefits from the
token management features provided by the \code{httr2} package,
including caching and automatic token refresh.

Since this function caches the token in memory, it will not persist across
sessions - this function will need to be run at the start of each session
to request a fresh token. Each token also expires after 1 hour, however
the automatic refresh behavior means that you will not need to run this
function multiple times per session (just once at the start).

For more information about Lightcast's authentication flow, see:
https://docs.lightcast.io/lightcast-api/docs/authentication-guide.
For more information about using \code{httr2} to manage authorization
using OAuth, see https://httr2.r-lib.org/articles/oauth.html.
}
\examples{
# Ensure API credentials have been saved to .Renviron before running
\dontrun{
authenticate_api()
}
# Simply running the function will retrieve the token and save it to an
# environment that persists until you end the R session.

}
